+++ framework/contracts/account/manager/tests/adapters.rs
@@ -531,7 +531,6 @@ fn subaccount_adapter_ownership() -> AResult {
         None,
         None,
         None,
-        None,
         &[],
     )?;