+++ framework/packages/abstract-adapter/src/lib.rs
@@ -169,6 +169,7 @@ pub mod mock {
     macro_rules! gen_adapter_mock {
     ($name:ident, $id:expr, $version:expr, $deps:expr) => {
         use $crate::std::adapter::*;
+        use $crate::sdk::base::Handler;
         use ::cosmwasm_std::Empty;
         use ::abstract_adapter::mock::{MockExecMsg, MockQueryMsg, MockInitMsg, MockAdapterContract, MockError};
         use ::cw_orch::environment::CwEnv;
@@ -216,6 +217,22 @@ pub mod mock {
 
         impl <T: ::cw_orch::prelude::CwEnv> ::abstract_interface::AdapterDeployer<T, MockInitMsg> for $name <T> {}
 
+        impl<Chain> $crate::abstract_interface::RegisteredModule for $name<Chain> {
+            type InitMsg = MockInitMsg;
+
+            fn module_id<'a>() -> &'a str {
+                $crate::traits::ModuleIdentification::module_id(&MOCK_ADAPTER)
+            }
+
+            fn module_version<'a>() -> &'a str {
+                MOCK_ADAPTER.version()
+            }
+
+            fn dependencies<'a>() -> &'a [$crate::objects::dependency::StaticDependency] {
+                MOCK_ADAPTER.dependencies()
+            }
+        }
+
         impl<T: ::cw_orch::prelude::CwEnv> Uploadable for $name<T> {
             fn wrapper() -> <Mock as ::cw_orch::environment::TxHandler>::ContractSource {
                 Box::new(ContractWrapper::<
@@ -252,6 +269,7 @@ pub mod mock {
     macro_rules! gen_adapter_old_mock {
     ($name:ident, $id:expr, $version:expr, $deps:expr) => {
         use $crate::std::adapter::*;
+        use $crate::sdk::base::Handler;
         use ::cosmwasm_std::Empty;
         use ::abstract_adapter::mock::{MockExecMsg, MockQueryMsg,  MockInitMsg, MockAdapterContract, MockError};
         use ::cw_orch::environment::CwEnv;
@@ -297,6 +315,22 @@ pub mod mock {
         pub struct $name ;
 
         impl ::abstract_interface::AdapterDeployer<::cw_orch::prelude::MockBech32, MockInitMsg> for $name <::cw_orch::prelude::MockBech32> {}
+
+        impl ::abstract_interface::RegisteredModule for $name<::cw_orch::prelude::MockBech32> {
+            type InitMsg = MockInitMsg;
+
+            fn module_id<'a>() -> &'a str {
+                $crate::traits::ModuleIdentification::module_id(&MOCK_ADAPTER)
+            }
+
+            fn module_version<'a>() -> &'a str {
+                MOCK_ADAPTER.version()
+            }
+
+            fn dependencies<'a>() -> &'a [$crate::objects::dependency::StaticDependency] {
+                MOCK_ADAPTER.dependencies()
+            }
+        }
 
         impl Uploadable for $name<::cw_orch::prelude::MockBech32> {
             fn wrapper() -> <MockBech32 as ::cw_orch::environment::TxHandler>::ContractSource {