+++ modules/contracts/apps/ping-pong/src/handlers/execute.rs
@@ -56,15 +56,15 @@ fn query_and_ping(
     module: App,
 ) -> AppResult {
     let ibc_client = module.ibc_client(deps.as_ref());
-    let dest_account_id = module
+    let remote_account_id = module
         .account_id(deps.as_ref())?
-        .into_dest_account_id(TruncatedChainId::new(env), opponent_chain.clone());
+        .into_remote_account_id(TruncatedChainId::new(env), opponent_chain.clone());
 
     let module_query = ibc_client.module_ibc_query(
         opponent_chain.clone(),
         InstalledModuleIdentification {
             module_info: module.module_info()?,
-            account_id: Some(dest_account_id),
+            account_id: Some(remote_account_id),
         },
         &crate::msg::QueryMsg::from(AppQueryMsg::BlockHeight {}),
         Callback::new(&PingPongCallbackMsg::QueryBlockHeight { opponent_chain })?,