+++ framework/docs/src/releases/CHANGELOG.md
@@ -5,18 +5,20 @@
 ### Added
 
 - `execute_with_funds` to Executor to attach funds to execution.
+- `stargate` feature for abstract-app, abstract-standalone and abstract-adapter packages.
 - New module type: `Service`, behaves the same as Native, but can be registered by any namespace.
 - `AbstractClient`: `service` to get api of Service module
 - `CustomExecuteHandler` To improve support for fully custom execute messages on Apps or Adapters
+- `balance` method for `AnsHost` to query balance of `AssetEntry`
 
 ### Changed
 
+- Ibc related renaming to add more consistency in namings
 - Account action on executor takes `impl IntoIter<Item = impl Into<AccountAction>>` instead of `Vec<AccountAction>`
 
 #### Abstract Client
 
 - `with_modules` method for Account Builder to add list of modules to install (`ModuleInstallConfig`)
-- `query_ans_balance` method for Account to query balance of `AssetEntry`
 - `query_module` method for Account to query given module on account without retrieving `Application` object
 - `module_installed` method for Account that returns `true` if module installed on account
 - `module_version_installed` method for Account that returns `true` if module of this version installed on account
@@ -24,10 +26,14 @@
 - `enable_ibc` added to Account builder.
 - `module_status` on AbstractClient that returns current status of the module.
 - `install_on_sub_account` now defaults to `false` in Account Builder
+- `Publisher` will check if dependencies of the module is registered in version control to the chain before publishing.
 
 ### Removed
 
 - Receive endpoints from abstract Modules
+- Value calculation logic from proxy contract.
+- `cw-semver` dependency removed
+
 
 ### Fixed